requirements.txt
beautifulsoup4==4.4.1
click==6.6
colorama==0.3.7
crypto==1.4.1
docopt==0.6.2
Flask==0.11.1
itsdangerous==0.24
Jinja2==2.8
MarkupSafe==0.23
Naked==0.1.31
nose==1.3.7
python-firebase==1.2
pythonds==1.0.1
PyYAML==3.11
requests==2.10.0
shellescape==3.4.1
termcolor==1.1.0
test==2.3.4.5
urllib2-file==0.2.1
virtualenv==15.0.1
Werkzeug==0.11.10
